created RelaySelectionDialog file
